% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrator.R
\name{calibrator}
\alias{calibrator}
\title{Calibrator HOBO data in CSV format}
\usage{
calibrator(list.data, columns = c(2, 7, 12), times, round = 7)
}
\arguments{
\item{list.data}{A list containing the HOBO CSV files.}

\item{columns}{The columns to be used for calibration.}

\item{times}{The times in a vector of dates to be included in the calibration process.}

\item{round}{The number of decimal places to round the results to.}
}
\value{
a data frame with the difrences for data correction, to use with corrector
}
\description{
Calculates the difference between HOBO devices under controlled conditions. This additive function calculates the difference among hobo loggers using a base correction to HOBO loggers.
}
\examples{
\donttest{
path ="/Desktop//calibration/files/"

file.exists(path)

folder=paste0(rep("canopy", 24), 1:24)

a=b=d=list()

for (i in seq_along(folder)){
  a[[i]] <- paste0(path, folder[i])
  b[[i]] <- hobinder(as.character(a[i]), channels = "ON" )
  d[[i]] <- hobocleaner(b[[i]], format = "ymd")
}

data.calibrated <- calibritor(calibrationfiles)
}
}
\author{
Ricardo I Alcala Briseno, \email{ria5282@psu.edu}
}
