% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correction.test.R
\name{correction.test}
\alias{correction.test}
\title{Correction Test for HOBO data from calibrator}
\usage{
correction.test(
  list.data,
  calibrationfile,
  w.var = c(3, 8, 13),
  times = times,
  threshold = c(1, 5, 10)
)
}
\arguments{
\item{list.data}{A list of CSV data frames containing the HOBO data.}

\item{calibrationfile}{A data frame representing the calibration file.}

\item{w.var}{A vector of column indices to be used in the correction.}

\item{times}{A vector of times for which the data is relevant.}

\item{threshold}{A vector of threshold values for passing the correction test.
The smaller the value the highest precision.}
}
\value{
A data frame with the differences for data correction, to be used
with a corrector.
}
\description{
This function calculates the difference among HOBO loggers, finding the
variance and using it as a base correction. It's designed to adjust HOBO
data based on calibration files and specified thresholds.
}
\examples{

path <- system.file("extdata/calibration", package = "hoboR")

folder=paste0(rep("canopy", 5), 1:5)

pathtoread = data = list()

for (i in seq_along(folder)){
  pathtoread[[i]] <- paste0(path, "/",folder[i])
  # Loading all hobo files
  data[[i]] <- hobinder(as.character(pathtoread[i]), header = TRUE, skip = 0,
   channels = "ON" ) # channels is a new feature
}
 
# Double-check you enter the same date format
times <- c("2022-03-22 01:00", "2022-03-22 02:00", "2022-03-22 03:00", 
            "2022-03-22 04:00","2022-03-22 05:00", "2022-03-22 06:00", 
            "2022-03-22 07:00", "2022-03-22 08:00","2022-03-22 09:00") 
            
variables <- c(3, 8, 13) # Select the weather variables 

meanvars <- calibrator(data, columns = variables, times = times)
 
correction.test(list.data = data, calibrationfile =  meanvars, 
                w.var = variables, 
                times = times, 
                threshold = c(1, 5, 10))
}
\author{
Ricardo I Alcala Briseno, \email{ria5282@psu.edu}
}
