% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_NAs.R
\name{count_NAs}
\alias{count_NAs}
\title{HOBO count NA's

This function counts the number of NA's in your data set}
\usage{
count_NAs(data, w.var)
}
\arguments{
\item{data}{dataframe with suspected NA's}

\item{w.var}{weather variables to test}
}
\value{
Returns the numbers of  NAs for the impossible values
}
\description{
HOBO software
}
\examples{
path <- system.file("extdata", package = "hoboR")

csvfiles <- hobinder(path, header = TRUE, skip = 1, channels = "OFF") 

cleancsv <- hobocleaner(csvfiles)

NAdata <- sensorfailures(cleancsv, condition = "<", threshold = c(0, 0), 
                            w.var = c("Rain", "Wetness"))
                            
count_NAs(NAdata, "Temp")

}
\author{
Ricardo I Alcala Briseno, \email{ria5282@psu.edu}
}
