% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ho.range.R
\name{ho.range}
\alias{ho.range}
\title{calculate date range temperature using hobo data
This function calculates hobo weather means for sampling rates
HOBO software}
\usage{
ho.range(data, start = "yyyy/mm/dd", end = "yyyy/mm/dd", round, na.rm = T)
}
\arguments{
\item{data}{a data frame with the hobo data and a \code{Date} column}

\item{start}{an early date in yyyy/mm/dd format}

\item{end}{a late date in yyyy/mm/dd format}
}
\value{
A data frame subset from start to end hobo range
}
\description{
calculate date range temperature using hobo data
This function calculates hobo weather means for sampling rates
HOBO software
}
\examples{
samples <- read.cvs(sampling.data)
site.ranges <- ho.range(samples, start = "2021/11/02",  end = "2021/11/23")
}
\author{
Ricardo I Alcala Briseno, \email{alcalabr@oregonstate.edu}
}
