% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hobinder.R
\name{hobinder}
\alias{hobinder}
\title{Reads HOBO data in CSV format}
\usage{
hobinder(path, channels = "OFF", ...)
}
\arguments{
\item{path}{select the path to the directory with the csv files}

\item{channels}{turn on or off additional channels in HOBO data logger, default "OFF"}

\item{...}{arguments to be passed to methods}
}
\value{
large csv file
}
\description{
Two functions that read the original data downloaded from HOBO software
adding file names as metadata for each .csv file and cleans the data
from duplicates creating a continuous file from all .csv's
}
\examples{

path <- system.file("extdata", package = "hoboR")

csvfiles <- hobinder(path, header = TRUE, skip = 1, channels = "OFF") 

head(csvfiles)

}
\author{
Ricardo I Alcala Briseno, \email{ria5282@psu.edu}
}
