% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hobotime.R
\name{hobotime}
\alias{hobotime}
\title{Summarise HOBO data by time intervals}
\usage{
hobotime(data, summariseby = "5 mins", na.rm = TRUE, na.action = na.omit)
}
\arguments{
\item{data}{a data frame with the hobo data and a \code{Date} column}

\item{summariseby}{a time interval in minmutes}

\item{na.rm}{logical vector TRUE or FALSE}

\item{na.action}{na.omit remove rows with NA's, na.pass keeps NA's}
}
\value{
a data frame summarized by minutes
}
\description{
This function calculates hobo weather by minutes
HOBO software
}
\examples{

path <- system.file("extdata", package = "hoboR")

csvfiles <- hobinder(path, header = TRUE, skip = 1, channels = "OFF") 

subset <- hobotime(csvfiles, summariseby = 5, 
                    na.rm = TRUE, na.action = na.pass)

head(subset)

}
\author{
Ricardo I Alcala Briseno, \email{ria5282@psu.edu}
}
