% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/horangedeprecated.R
\name{horange}
\alias{horange}
\title{Subset data within a date range}
\usage{
horange(data, start = "yyyy/mm/dd", end = "yyyy/mm/dd", round, na.rm = T)
}
\arguments{
\item{data}{A data frame with the HOBO data and a \code{Date} column}

\item{start}{Start date for the subset (in "yyyy/mm/dd" format)}

\item{end}{End date for the subset (in "yyyy/mm/dd" format)}

\item{round}{Logical, whether to round the start and end dates to the nearest available dates in the data}

\item{na.rm}{Logical, whether to remove NAs from the result}
}
\value{
A subset of the original data frame within the specified date range
}
\description{
This function subsets data within a specified date range.
HOBO software
}
\examples{
cleaned <- hobocleaner(data, format = "yymd")
site.ranges <- horange(cleaned, start = "2021/11/02",  end = "2021/11/23")
}
\author{
Ricardo I Alcala Briseno, \email{alcalabr@oregonstate.edu}
}
