% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NAsensorfailures.R
\name{NAsensorfailures}
\alias{NAsensorfailures}
\title{HOBO Remove Values}
\usage{
NAsensorfailures(
  data,
  condition = ">",
  threshold = c(34, 8),
  opt = c("Temp", "Rain")
)
}
\arguments{
\item{data}{Remove impossible values from HOBO data frame from \verb{original csv}
OR \code{hobocleaner} OR \code{hobotime}}

\item{condition}{The condition for removal, one of (">", "<", "==", ">=", "<=")}

\item{threshold}{Numeric vector specifying the threshold values for removal}

\item{opt}{Select the header to remvoe impossible values}
}
\value{
Returns the data with NAs for the impossible values
}
\description{
Functions that remove impossible values
HOBO software
}
\examples{
data <- hobunder(loadAllcsvs)
data_clean <- hobocleaner(data)
NAsensorfailures(data, condition = ">", threshold = c(50, 3000, 101), opt = c("Temp", "Rain", "Wetness"))
NAdata <- NAsensorfailures(data, condition = "<", threshold = c(0, 0), opt = c("Rain", "Wetness"))
}
\author{
Ricardo I Alcala Briseno, \email{alcalabr@oregonstate.edu}
}
