% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.rates.R
\name{sampling.rates}
\alias{sampling.rates}
\title{Calculates sampling rates from a CSV format}
\usage{
sampling.rates(samples, n, round)
}
\arguments{
\item{samples}{a csv with the format}

\item{n}{Mandatory. Specifies the number of replicates of the experiment}

\item{round}{Optional. Specifies the number of decimal places for rounding
the output incidence calculated from a csv table designed for baiting
Phytophthora}
}
\value{
smaller data frame with incidence and rates
}
\description{
This function calculates incidence and rates for baiting
Phytophthora collected on dates for baited and removed leaves
HOBO software
}
\examples{

path <- system.file("extdata/sampling", package = "hoboR")
#' 
samples <- read.csv(paste0(path, "/", "sampling.csv"))
 
samp.rates <- sampling.rates(samples, n = 9, round = 2)

}
\author{
Ricardo I Alcala Briseno, \email{ria5282@psu.edu}
}
