% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.trends.R
\name{sampling.trends}
\alias{sampling.trends}
\title{calcula sampling bate rates temperature using weather data and
sampling rates for data collection
This function calculates hobo weather means for sampling rates
Ex: phytophthora collected on dates for baited and removed leaves
HOBO software}
\usage{
sampling.trends(hobomeans, samp.rates, round)
}
\arguments{
\item{samples}{a csv with the format
incidence calculated from a csv table designed for baiting Phytophtora}
}
\value{
smaller data frame with incidence and rates
}
\description{
calcula sampling bate rates temperature using weather data and
sampling rates for data collection
This function calculates hobo weather means for sampling rates
Ex: phytophthora collected on dates for baited and removed leaves
HOBO software
}
\examples{
samples <- read.cv(sampling.data)
samp.rates <- samplingrates(samples)
}
\author{
Ricardo I Alcala Briseno, \email{alcalabr@oregonstate.edu}
}
