% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplingrates.R
\name{samplingrates}
\alias{samplingrates}
\title{calcula sampling rates from a CSV format}
\usage{
samplingrates(samples, n, round)
}
\arguments{
\item{samples}{a csv with the format
incidence calculated from a csv table designed for baiting Phytophtora}
}
\value{
smaller data frame with incidence and rates
}
\description{
This function calculates incidence and rates for baiting
phytophthora collected on dates for baited and removed leaves
HOBO software
}
\examples{
samples <- read.cv(sampling.data)
samp.rates <- samplingrates(samples)
}
\author{
Ricardo I Alcala Briseno, \email{alcalabr@oregonstate.edu}
}
