% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testhobolist.R
\name{testhobolist}
\alias{testhobolist}
\title{Test HOBO data for calibration}
\usage{
testhobolist(data, times)
}
\arguments{
\item{data}{a list of CVS data containing the hobo}

\item{times}{a series of times <- c("2022-03-22 01:00", "2022-03-22 02:00", "2022-03-22 03:00")}
}
\value{
a data frame with the total entries and the count of entries present in each data set
}
\description{
Check if the times date are present in the list of HOBO files
}
\examples{
\dontrun{
data <- testhobolist(data, times)
}
}
\author{
Ricardo I Alcala Briseno, \email{ria5282@psu.edu}
}
