% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrator.R
\docType{data}
\name{times}
\alias{times}
\title{mean HOBO data in CSV format}
\format{
An object of class \code{character} of length 9.
}
\usage{
times
}
\arguments{
\item{data}{}
}
\value{
calibrated data frame
}
\description{
Functions that calculate the parameter to correct with calibrated data from the
HOBO software
}
\examples{
calibratedcsv <- calibritor(calibration-data)
}
\author{
Ricardo I Alcala Briseno, \email{alcalabr@oregonstate.edu}
}
\keyword{datasets}
