% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrator.R
\name{calibrator}
\alias{calibrator}
\title{Calibrator HOBO data in CSV format}
\usage{
calibrator(list.data, columns = c(2, 7, 12), times, round = 7)
}
\arguments{
\item{list.data}{A list containing the HOBO CSV files.}

\item{columns}{The columns to be used for calibration.}

\item{times}{The times in a vector of dates to be included in the calibration
process.}

\item{round}{The number of decimal places to round the results to.}
}
\value{
a data frame with the differences for data correction, to use with
corrector
}
\description{
Calculates the difference between HOBO devices under controlled conditions.
This additive function calculates the difference among hobo loggers using a
base correction to HOBO loggers.
}
\examples{

path <- system.file("extdata/calibration", package = "hoboR")
 
folder=paste0(rep("canopy", 5), 1:5)

pathtoread = data = list()

for (i in seq_along(folder)){
  pathtoread[[i]] <- paste0(path, "/",folder[i])
  # Loading all hobo files
  data[[i]] <- hobinder(as.character(pathtoread[i]), header = TRUE, skip = 0,
   channels = "ON" ) # channels is a new feature
}

# Make sure you enter the date & time format with zeros, 
# for example 08:00 instead of 8:00 for 8am.
times <- c("2022-03-22 01:00", "2022-03-22 02:00", "2022-03-22 03:00", 
           "2022-03-22 04:00", "2022-03-22 05:00", "2022-03-22 06:00", 
           "2022-03-22 07:00", "2022-03-22 08:00", "2022-03-22 09:00") 
           
variables <- c(3, 8, 13) # Select the weather variables 
           
calibrationmeans <- calibrator(data, columns= variables, times = times) 

}
\author{
Ricardo I Alcala Briseno, \email{ria5282@psu.edu}
}
