% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrator.R
\name{calibrator}
\alias{calibrator}
\title{mean HOBO data in CSV format}
\usage{
calibrator(data, formula = "y = a + b", times = c(12:0, 13:0, 14:0))
}
\arguments{
\item{data}{}
}
\value{
calibrated data frame
}
\description{
Functions that calculate the parameter to correct with calibrated data from the
HOBO software
}
\examples{
calibratedcsv <- calibritor(calibration-data)
}
\author{
Ricardo I Alcala Briseno, \email{alcalabr@oregonstate.edu}
}
