% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrator.R
\name{calibrator}
\alias{calibrator}
\title{Calibrator HOBO data in CSV format}
\usage{
calibrator(list.data = data, columns = c(2, 7, 12), times = times, round = 7)
}
\arguments{
\item{a}{list containing the HOBO csv files}

\item{select}{the times in a vector of dates to be included in the calibration process}
}
\value{
a data frame with the difrences for data correction, to use with corrector
}
\description{
Calculates the difference between HOBO devices under controlled conditions. This additive function calculates the difference among hobo loggers using a base correction to HOBO loggers.
}
\examples{
path = "~/Desktop/testsky/calibration/originalfiles/"
file.exists(path)
folder=paste0(rep("canopy", 24), 1:24)
a=b=d=list()
for (i in seq_along(folder)){
 a[[i]] <- paste0(path, folder[i])
 b[[i]] <- hobinder(as.character(a[i]), channels = "ON" )
 d[[i]] <- hobocleaner(b[[i]], format = "ymd")
}
data.calibrated <- calibritor(calibrationfiles)
}
\author{
Ricardo I Alcala Briseno, \email{alcalabr@oregonstate.edu}
}
