% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrator.R
\name{calibrator}
\alias{calibrator}
\title{Calibrator HOBO data in CSV format}
\usage{
calibrator(
  list.data = data,
  formula = "y = a + b",
  columns = c(2, 7, 12),
  times = times,
  round = 7
)
}
\arguments{
\item{times}{a series of times <- c("2022-03-22 01:00", "2022-03-22 02:00", "2022-03-22 03:00")}

\item{data}{a list of CVS data containing the hobo}
}
\value{
a data frame with the difrences for data correction, to use with corrector
}
\description{
Additive function to calculate the difference among hobo loggers to calibrate
using a base correction to the data
}
\examples{
path = "~/Desktop/testsky/calibration/originalfiles/"
file.exists(path)
folder=paste0(rep("canopy", 24), 1:24)
pathtoread=calibrationfiles=hobocleaned=data=list()
for (i in seq_along(folder)){
 pathtoread[[i]] <- paste0(path, folder[i])
 calibrationfiles[[i]] <- hobinder(as.character(pathtoread[i]), channels = "ON" ) # channels is a new feature
 data[[i]]=hobocleaned[[i]] <- hobocleaner(calibrationfiles[[i]], format = "ymd")
}
data.calibrated <- calibritor(calibrationfiles)

}
\author{
Ricardo I Alcala Briseno, \email{alcalabr@oregonstate.edu}
}
