% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hobocorrection.R
\name{correction}
\alias{correction}
\title{Correction test for HOBO data from calibrator}
\usage{
correction(data, w.var = "FULL", calibrate = calibrate)
}
\arguments{
\item{data}{a list of CVS data containing the hobo}

\item{w.var}{a column to correct the weather variable e.g., Temperature, RH (relative humidity), or FULL, will use the output of calibrator}

\item{calibrate}{a value to correct the weather variable, must be numeric or USEFILE, will use the output of calibrator}
}
\value{
a data frame with the differences for data correction, to use with corrector
}
\description{
Additive function to calculate the difference among hobo loggers to calibrate
using a base correction to the data
}
\examples{

path <- system.file("extdata/calibration", package = "hoboR")

folder=paste0(rep("canopy", 5), 1:5)

pathtoread = data = list()

for (i in seq_along(folder)){
  pathtoread[[i]] <- paste0(path, "/",folder[i])
  # Loading all hobo files
  data[[i]] <- hobinder(as.character(pathtoread[i]), header = TRUE, skip = 0,
   channels = "ON" ) # channels is a new feature
}
 
# Double-check you enter the same date format
times <- c("2022-03-22 01:00", "2022-03-22 02:00", "2022-03-22 03:00", 
            "2022-03-22 04:00","2022-03-22 05:00", "2022-03-22 06:00", 
            "2022-03-22 07:00", "2022-03-22 08:00","2022-03-22 09:00") 
            
variables <- c(3, 8, 13) # Select the weather variables 

meanvars <- calibrator(data, columns = variables, times = times)
                
calibratedfiles <- correction(data = data, 
                              w.var = "FULL", 
                              calibrate = meanvars)

}
\author{
Ricardo I Alcala Briseno, \email{ria5282@psu.edu}
}
