% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hobocorrection.R
\name{correction}
\alias{correction}
\title{Correction test for HOBO data from calibrator}
\usage{
correction(data, w.var = "FULL", calibrate = calibrate)
}
\arguments{
\item{data}{a list of CVS data containing the hobo}

\item{w.var}{a column to correct the weather variable e.g., Temperature, RH (relative humidity), or FULL, will use the output of calibrator}

\item{calibrate}{a value to correct the weather variable, must be numeric or USEFILE, will use the output of calibrator}
}
\value{
a data frame with the differences for data correction, to use with corrector
}
\description{
Additive function to calculate the difference among hobo loggers to calibrate
using a base correction to the data
}
\examples{
\dontrun{
path = "~/Desktop/testsky/calibration/originalfiles/"

filestocorrect <- calibritor(calibrationfiles)

calibratedfiles <- correction(data, calibration, calibrate=)
}
}
\author{
Ricardo I Alcala Briseno, \email{ria5282@psu.edu}
}
