% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hobor.correction.R
\name{correction}
\alias{correction}
\title{Correction test for HOBO data from calibrator}
\usage{
correction(data)
}
\arguments{
\item{data}{a list of CVS data containing the hobo}

\item{times}{a series of times <- c("2022-03-22 01:00", "2022-03-22 02:00", "2022-03-22 03:00")}
}
\value{
a data frame with the difrences for data correction, to use with corrector
}
\description{
Additive function to calculate the difference among hobo loggers to calibrate
using a base correction to the data
}
\examples{
path = "~/Desktop/testsky/calibration/originalfiles/"
filestocorrect <- calibritor(calibrationfiles)
calibratedfiles <- correction(data, calibration, calibrate=)
}
\author{
Ricardo I Alcala Briseno, \email{alcalabr@oregonstate.edu}
}
