% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correction.test.R
\name{correction.test}
\alias{correction.test}
\title{Correction test for HOBO data from calibrator}
\usage{
correction.test(
  list.data = data,
  calibrationfile = x,
  columns = c(2, 7, 12),
  times = times,
  threshold = 1
)
}
\arguments{
\item{times}{a series of times <- c("2022-03-22 01:00", "2022-03-22 02:00", "2022-03-22 03:00")}

\item{data}{a list of CVS data containing the hobo}
}
\value{
a data frame with the difrences for data correction, to use with corrector
}
\description{
Additive function to calculate the difference among hobo loggers to calibrate
using a base correction to the data
}
\examples{
path = "~/Desktop/testsky/calibration/originalfiles/"
data.calibrated <- calibritor(calibrationfiles)
}
\author{
Ricardo I Alcala Briseno, \email{alcalabr@oregonstate.edu}
}
