% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_NAs.R
\name{count_NAs}
\alias{count_NAs}
\title{HOBO count NA's

This function counts the number of NA's in your data set}
\usage{
count_NAs(data, w.var)
}
\arguments{
\item{data}{dataframe with suspected NA's}

\item{w.var}{weather variables to test}
}
\value{
Returns the numbers of  NAs for the impossible values
}
\description{
HOBO software
}
\examples{
\dontrun{
df <- hobocleaner(df)

NAdata <- sensorfailures(data, condition = "<", threshold = c(0, 0), w.var = c("Rain", "Wetness"))

count_NAs(NAdata, "Temp")
}
}
\author{
Ricardo I Alcala Briseno, \email{alcalabr@oregonstate.edu}
}
