% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hoboR.R
\name{hobinder}
\alias{hobinder}
\title{Reads HOBO data in CSV format}
\usage{
hobinder(path, col.names = col.names)
}
\arguments{
\item{path}{to indicate the path to the csv files}

\item{files}{get the bind csv file, remove duplicates and merge the data}
}
\value{
large csv file
}
\description{
Two functions that read the original data downloaded from HOBO software
adding file names as metadata for each .csv file and cleans the data
from duplicates creating a continuous file from all .csv's
}
\examples{
path_to_csvs <- '~mydirectory/myfiles.csv/'
loadAllcsvs <- hobinder(path_to_csvs)
finalcsv <- hobocleaner(loadAllcsvs)
}
\author{
Ricardo I Alcala Briseno, \email{alcalabr@oregonstate.edu}
}
