% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hobocleaner.R
\name{hobocleaner}
\alias{hobocleaner}
\title{Reads HOBO data in CSV format}
\usage{
hobocleaner(file, format = "ymd", na.rm = T)
}
\arguments{
\item{file}{CSV from \code{hobinder}}

\item{format}{Select the time format, month, day, and year (mdy), year, month, and day (ymd) or
year two digits, month and day (yymd)}

\item{na.rm}{TRUE or FALSE to remove NAs, TRUE is default}
}
\value{
formatted data frame and duplicate values removed
}
\description{
Functions that cleans the original data downloaded from the HOBO software,
cleans the data and removes duplicates creating a continuous file from all .csv's
}
\examples{
\dontrun{
path_to_csvs <- "path/to/hobo.csv"

csvs <- hobinder(path_to_csvs)

cleancsvs <- hobocleaner(csvs)
}
}
\author{
Ricardo I Alcala Briseno, \email{alcalabr@oregonstate.edu}
}
