% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hobocleaner-deprecated.R, R/hobocleaner.R
\name{hobocleaner}
\alias{hobocleaner}
\title{Reads HOBO data in CSV format}
\usage{
hobocleaner(file, format = "ymd", na.rm = T)

hobocleaner(file, format = "ymd", na.rm = T)
}
\arguments{
\item{file}{Get the binded CSV file from \code{hobinder} and remove duplicates to merge the data.}

\item{format}{Choose between month, day, and year (mdy) or year, month, and day (ymd) or
year two digits, month and day (yymd)}

\item{na.rm}{TRUE or FALSE to remove NAs, TRUE is default}
}
\value{
large clean csv file

#' @importFrom lubridate ymd_hms
#' @importFrom dplyr

data frame without duplicate values
}
\description{
Functions that cleans the original data downloaded from the HOBO software
cleans the data and removes duplicates creating a continuous file from all .csv's

Functions that cleans the original data downloaded from the HOBO software,
cleans the data and removes duplicates creating a continuous file from all .csv's
}
\examples{
path_to_csvs <- '~mydirectory/myfiles.csv/'
loadAllcsvs <- hobinder(path_to_csvs)
finalcsv <- hobocleaner(loadAllcsvs)
path_to_csvs <- "~/Documents/"
loadAllcsvs <- hobinder(path_to_csvs)
file <- hobocleaner(loadAllcsvs)
}
\author{
Ricardo I Alcala Briseno, \email{alcalabr@oregonstate.edu}
}
