% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hobocleaner.R
\name{hobocleaner}
\alias{hobocleaner}
\title{Reads HOBO data in CSV format}
\usage{
hobocleaner(file)
}
\arguments{
\item{file}{get the binded csv file from \code{hobinder}, and remove duplicates and merge the data}
}
\value{
large clean csv file

#' @importFrom lubridate ymd_hms
#' @importFrom dplyr
}
\description{
Functions that cleans the original data downloaded from the HOBO software
cleans the data and removes duplicates creating a continuous file from all .csv's
}
\examples{
path_to_csvs <- '~mydirectory/myfiles.csv/'
loadAllcsvs <- hobinder(path_to_csvs)
finalcsv <- hobocleaner(loadAllcsvs)
}
\author{
Ricardo I Alcala Briseno, \email{alcalabr@oregonstate.edu}
}
