% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hobocorrelations.R
\name{hobocorrelations}
\alias{hobocorrelations}
\title{Correlation plot for variables within a time range}
\arguments{
\item{data}{A data frame with the HOBO data and a \code{Date} column}

\item{summariseby}{Provide the interval date to present (e.g., "month")}

\item{by}{Summary function for aggregation (e.g., "mean")}

\item{na.rm}{Logical, whether to remove NAs from the result}
}
\value{
A ggplot object representing the correlation heatmap
}
\description{
This function provides a time point for a specified number of days.
HOBO software
}
\examples{

path <- system.file("extdata", package = "hoboR")

csvfiles <- hobinder(path, header = TRUE, skip = 1, channels = "OFF") 

cleaned <- hobocleaner(csvfiles, format = "ymd")

hobocorrelations(cleaned, summariseby = "month", by = "mean", na.rm = FALSE)

}
\author{
Ricardo I Alcala Briseno, \email{ria5282@psu.edu}
}
