% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hoborange.R
\name{hoborange}
\alias{hoborange}
\title{Calculate Date Range Temperature Using HOBO Data}
\usage{
hoborange(
  data,
  start = "2022-08-04 09:05",
  end = "2022-10-04 09:05",
  na.rm = TRUE
)
}
\arguments{
\item{data}{A data frame containing the HOBO data, including a \code{Date} column in POSIXct format.}

\item{start}{The start of the date range in "yyyy-mm-dd HH:MM" format.}

\item{end}{The end of the date range in "yyyy-mm-dd HH:MM" format.}

\item{na.rm}{A logical value indicating whether NA values should be removed before calculation.}
}
\value{
A subset of the original data frame limited to the specified date range.
}
\description{
This function calculates the means for weather data collected by HOBO loggers
over a specified date range. It's designed for use with data exported from HOBO software.
}
\examples{

path <- system.file("extdata", package = "hoboR")

csvfiles <- hobinder(path, header = TRUE, skip = 1, channels = "OFF") 

cleancsv <- hobocleaner(csvfiles)

site.ranges <- hoborange(cleancsv, start = "2022-08-04 09:05",  
                          end = "2022-10-04 09:05")

}
\author{
Ricardo I Alcala Briseno, \email{ria5282@psu.edu}
}
