% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hoborange.R
\name{hoborange}
\alias{hoborange}
\title{Calculate Date Range Temperature Using HOBO Data}
\usage{
hoborange(
  data,
  start = "1910-09-16 06:00",
  end = "1920-12-01 12:00",
  na.rm = T
)
}
\arguments{
\item{data}{A data frame containing the HOBO data, including a \code{Date} column in POSIXct format.}

\item{start}{The start of the date range in "yyyy-mm-dd HH:MM" format.}

\item{end}{The end of the date range in "yyyy-mm-dd HH:MM" format.}

\item{na.rm}{A logical value indicating whether NA values should be removed before calculation.}

\item{round}{Logical indicating whether to round the results (not used in the example, consider removing or implementing).}
}
\value{
A subset of the original data frame limited to the specified date range.
}
\description{
This function calculates the means for weather data collected by HOBO loggers
over a specified date range. It's designed for use with data exported from HOBO software.
}
\examples{
samples <- read.cvs(sampling.data)
site.ranges <- ho.range(samples, start = "1910/09/16",  end = "1920/12/01")
}
\author{
Ricardo I Alcala Briseno, \email{alcalabr@oregonstate.edu}
}
