% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hobotime.R
\name{hobotime}
\alias{hobotime}
\title{Summarise HOBO data by time intervals}
\usage{
hobotime(data, summariseby = "5 mins", na.rm = T)
}
\arguments{
\item{data}{a data frame with the hobo data and a \code{Date} column}

\item{summariseby}{a time interval in minmutes}

\item{na.rm}{logical vector TRUE or FALSE}
}
\value{
a data frame summarised by minutes
}
\description{
This function calculates hobo weather by minutes
HOBO software
}
\examples{
hobocleaned <- hobocleaner(files, format = "ymd")

hobosubset <- hobotime(cleaned, summariseby = 5)
}
\author{
Ricardo I Alcala Briseno, \email{alcalabr@oregonstate.edu}
}
