% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hobotime.R
\name{hobotime}
\alias{hobotime}
\title{Summarise HOBO data by time intervals}
\arguments{
\item{data}{a data frame with the hobo data and a \code{Date} column}

\item{summariseby}{a time interval in minmutes}

\item{na.rm}{logical vector TRUE or FALSE}

\item{na.action}{na.omit remove rows with NA's, na.pass keeps NA's}

\item{...}{arguments to be passed to methods}
}
\value{
a data frame summarized by minutes
}
\description{
This function calculates hobo weather by minutes
HOBO software
}
\examples{
\dontrun{
hobocleaned <- hobocleaner(files, format = "ymd")

hobosubset <- hobotime(cleaned, summariseby = 5, na.rm = TRUE, na.action = na.pass)
}
}
\author{
Ricardo I Alcala Briseno, \email{alcalabr@oregonstate.edu}
}
