% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impossiblevalues.R
\name{impossiblevalues}
\alias{impossiblevalues}
\title{HOBO impossible values}
\usage{
impossiblevalues(data, showrows = 10, ...)
}
\arguments{
\item{data}{Cleaned hobo data frame from \verb{original csv} or \code{hobocleaner} and \code{hobotime}}

\item{showrows}{Number of rows to show for maximum values, default is 10}

\item{...}{arguments to be passed to methods}
}
\value{
Gives the rows with impossible values
}
\description{
Functions that gets the mean by date of the cleaned data downloaded from the
HOBO software
}
\examples{
data <- hobocleaner(loadAllcsvs)

impossiblevalues(data, showrows = 10)

}
\author{
Ricardo I Alcala Briseno, \email{alcalabr@oregonstate.edu}
}
