% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanhobo2.R
\name{meanhobo}
\alias{meanhobo}
\title{mean HOBO data in CSV format}
\arguments{
\item{data}{cleaned hobo data frame from \code{hobocleaner}i}

\item{summarisedby}{select a time interval 60 min, 24 hours, 1 day}

\item{na.rm}{TRUE or FALSE to remove NAs, TRUE is default}

\item{minmax}{TRUE or FALSE to retain min and max temperatures}
}
\value{
smaller data frame with means and standard deviation
}
\description{
Functions that gets the summary statistics by summarizing by date the cleaned data downloaded from the HOBO software
}
\examples{

path <- system.file("extdata", package = "hoboR")

data <- hobinder(path, header = TRUE, skip = 1, channels = "OFF") 

cleandata <- hobocleaner(data, format = "ymd")
 
# this function needs clean data#needs clean data 
hobodata <- meanhobo(cleandata, summariseby = "5 mins", 
                                    na.rm = TRUE, minmax = TRUE)

}
\author{
Ricardo I Alcala Briseno, \email{ria5282@psu.edu}
}
