% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanhobo.R
\name{meanhobo}
\alias{meanhobo}
\title{mean HOBO data in CSV format}
\usage{
meanhobo(cleandata)
}
\arguments{
\item{mean}{and sd of the cleaned data frame \code{hobocleaner}}
}
\value{
smaller data frame with means and standar deviation
}
\description{
Functions that gets the mean by date of the cleaned data downloaded from the
HOBO software
}
\examples{
cleanedcsv <- hobocleaner(loadAllcsvs)
mean.vars <- meanhobo(cleanedcsv)
}
\author{
Ricardo I Alcala Briseno, \email{alcalabr@oregonstate.edu}
}
