% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanhobo2.R
\name{meanhobo}
\alias{meanhobo}
\title{mean HOBO data in CSV format}
\arguments{
\item{data}{cleaned hobo data frame from \code{hobocleaner}i}

\item{summarisedby}{select a time interval 60 min, 24 hours, 1 day}

\item{na.rm}{TRUE or FALSE to remove NAs, TRUE is default}

\item{minmax}{TRUE or FALSE to retain min and max temperatures}
}
\value{
smaller data frame with means and standard deviation
}
\description{
Functions that gets the summary statistics by summarizing by date the cleaned data downloaded from the HOBO software
}
\examples{
\dontrun{
hoboclean <- hobocleaner(loadAllcsvs)

hobomeans <- meanhobo(cleanedcsv)
}
}
\author{
Ricardo I Alcala Briseno, \email{alcalabr@oregonstate.edu}
}
