% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanhobo2.R
\name{meanhobo}
\alias{meanhobo}
\title{mean HOBO data in CSV format}
\usage{
meanhobo(data, summariseby = "24 hours", na.rm = T, minmax = F)
}
\arguments{
\item{data}{cleaned hobo data frame from \code{hobocleaner}}

\item{na.rm}{TRUE or FALSE to remove NAs, TRUE is default}

\item{minmax}{TRUE or FALSE to retain min and max temperatures}
}
\value{
smaller data frame with means and standard deviation
}
\description{
Functions that gets the summary statistics by summarizing by date the cleaned data downloaded from the HOBO software
}
\examples{
hoboclean <- hobocleaner(loadAllcsvs)

hobomeans <- meanhobo(cleanedcsv)
}
\author{
Ricardo I Alcala Briseno, \email{alcalabr@oregonstate.edu}
}
