% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanhobo-deprecated.R, R/meanhobo.R
\name{meanhobo}
\alias{meanhobo}
\title{mean HOBO data in CSV format}
\usage{
meanhobo(data, summariseby = "24 hours", na.rm = T)

meanhobo(data, summariseby = "24 hours", na.rm = T)
}
\arguments{
\item{data}{cleaned hobo data frame from \code{hobocleaner}}

\item{na.rm}{TRUE or FALSE to remove NAs, TRUE is default}

\item{mean}{and sd of the cleaned data frame \code{hobocleaner}}
}
\value{
smaller data frame with means and standar deviation

smaller data frame with means and standard deviation
}
\description{
Functions that gets the mean by date of the cleaned data downloaded from the
HOBO software

Functions that gets the mean by date of the cleaned data downloaded from the
HOBO software
}
\examples{
cleanedcsv <- hobocleaner(loadAllcsvs)
mean.vars <- meanhobo(cleanedcsv)
cleanedcsv <- hobocleaner(loadAllcsvs)
mean.vars <- meanhobo(cleanedcsv)
}
\author{
Ricardo I Alcala Briseno, \email{alcalabr@oregonstate.edu}
}
