% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.trends.R
\name{sampling.trends}
\alias{sampling.trends}
\title{Calculate the sampling rates using the baiting incidence and weather data
This function calculates hobo weather means for sampling rates
Ex: phytophthora collected on dates for baited and removed leaves}
\usage{
sampling.trends(hobomeans, samp.rates, round, na.rm = T)
}
\arguments{
\item{hobomeans}{a data frame with the hobo means}

\item{samp.rates}{a data frame with incidence summary of collected dates}

\item{round}{Optional. Specifies the number of decimal places for rounding the output.}

\item{na.rm}{remove NA's from data}
}
\value{
A data frame with summarized weather variables by data and location,
as incidence rates
}
\description{
Calculate the sampling rates using the baiting incidence and weather data
This function calculates hobo weather means for sampling rates
Ex: phytophthora collected on dates for baited and removed leaves
}
\examples{
samples <- read.cv(sampling.data)

hobomeans <- meanhobo(hobocleaned)

samp.rates <- samplingrates(samples, n = 9, round = 2)

Site <- sampling.trends(hobomeans, samp.rates, round = 2)
}
\author{
Ricardo I Alcala Briseno, \email{alcalabr@oregonstate.edu}
}
