% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.trends.R
\name{sampling.trends}
\alias{sampling.trends}
\title{calcula sampling bate rates temperature using weather data and
sampling rates for data collection
This function calculates hobo weather means for sampling rates
Ex: phytophthora collected on dates for baited and removed leaves
HOBO software}
\usage{
sampling.trends(hobomeans, samp.rates, round, na.rm = T)
}
\arguments{
\item{samp.rates}{a data frame with incidence summary of collected dates
for baiting Phytophtora}

\item{hobodata}{a data frame with the hobo means}
}
\value{
A data frame with summarized weather variables by data and location,
as incidence rates
}
\description{
calcula sampling bate rates temperature using weather data and
sampling rates for data collection
This function calculates hobo weather means for sampling rates
Ex: phytophthora collected on dates for baited and removed leaves
HOBO software
}
\examples{
samples <- read.cv(sampling.data)
hobomeans <- meanhobo(hobocleaned)
samp.rates <- samplingrates(samples, n = 9, round = 2)
Site <- sampling.trends(hobomeans, samp.rates, round = 2)
}
\author{
Ricardo I Alcala Briseno, \email{alcalabr@oregonstate.edu}
}
