% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testhobolist.R
\name{testhobolist}
\alias{testhobolist}
\title{Test HOBO data for calibration}
\usage{
testhobolist(data, times)
}
\arguments{
\item{data}{a list of CVS data containing hobo data}

\item{times}{a series of times <- c("2022-03-22 01:00", "2022-03-22 02:00",
"2022-03-22 03:00")}
}
\value{
a data frame with the total entries and the count of entries present
in each data set
}
\description{
Check if the times date are present in the list of HOBO files
}
\examples{

path <- system.file("extdata/calibration", package = "hoboR")
 
folder=paste0(rep("canopy", 5), 1:5)

pathtoread = data = list()

for (i in seq_along(folder)){
  pathtoread[[i]] <- paste0(path, "/",folder[i])
  # Loading all hobo files
  data[[i]] <- hobinder(as.character(pathtoread[i]), header = TRUE, skip = 0,
   channels = "ON" ) # channels is a new feature
}

# Make sure you enter the date & time format with zeros, 
# for example 08:00 instead of 8:00 for 8am.
times <- c("2022-03-22 01:00", "2022-03-22 02:00", "2022-03-22 03:00", 
           "2022-03-22 04:00", "2022-03-22 05:00", "2022-03-22 06:00", 
           "2022-03-22 07:00", "2022-03-22 08:00", "2022-03-22 09:00") 

data <- testhobolist(data, times)

}
\author{
Ricardo I Alcala Briseno, \email{ria5282@psu.edu}
}
