% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timestamp.R
\name{timestamp}
\alias{timestamp}
\title{Timestamp for Specific Intervals}
\usage{
timestamp(
  data,
  stamp = "yyyy/mm/dd: ss",
  by = "24 hours",
  days = 100,
  na.rm = TRUE,
  plot = TRUE,
  var = "Temp"
)
}
\arguments{
\item{data}{A data frame with the HOBO data and a \code{Date} column}

\item{stamp}{Provide a date}

\item{by}{Provide the interval date to present (e.g., "24 hours")}

\item{days}{Number of days for the interval}

\item{na.rm}{Logical, whether to remove NAs from the result}

\item{plot}{Logical, whether to generate a plot}

\item{var}{Variable to plot (default is "Temp")}
}
\value{
A data frame summarized by minutes
}
\description{
This function provides a time point for a specified number of days.
HOBO software
}
\examples{

path <- system.file("extdata", package = "hoboR")

data <- hobinder(path, header = TRUE, skip = 1, channels = "OFF")

cleandata <- hobocleaner(data, format = "ymd")

datastamp <- timestamp(cleandata, stamp ="2022-08-04 12:00", 
                        by = "1 hour", days = 60, na.rm = TRUE, 
                        plot = TRUE, var = "Temp")

}
\author{
Ricardo I Alcala Briseno, \email{ria5282@psu.edu}
}
