% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timestamp.R
\name{timestamp}
\alias{timestamp}
\title{Timestamp for Specific Intervals}
\arguments{
\item{data}{A data frame with the HOBO data and a \code{Date} column}

\item{stamp}{Provide a date}

\item{by}{Provide the interval date to present (e.g., "24 hours")}

\item{days}{Number of days for the interval}

\item{na.rm}{Logical, whether to remove NAs from the result}

\item{plot}{Logical, whether to generate a plot}

\item{var}{Variable to plot (default is "Temp")}
}
\value{
A data frame summarized by minutes
}
\description{
This function provides a time point for a specified number of days.
HOBO software
}
\examples{
files <- hobinder(path)

cleaned <- hobocleaner(files, format = "ymd")

timestamp <- timestamp(cleaned, stamp = "yyyy/mm/dd: ss", by = "24 hours", days = 100, na.rm = TRUE, plot = TRUE, var = "Temp")
}
\author{
Ricardo I Alcala Briseno, \email{alcalabr@oregonstate.edu}
}
